+++ src/helpers/randomElementOfArray.ts
@@ -6,7 +6,7 @@ import { randomInt } from "./randomInt.js";
  * @param array The array
  * @returns A random element of the array.
  */
-export function randomElementOfArray<T>(array: NonEmptyArray<T>): T;
+export function randomElementOfArray<T>(array: Readonly<NonEmptyArray<T>>): T;
 
 /**
  * Returns a random element from an array.
@@ -14,8 +14,8 @@ export function randomElementOfArray<T>(array: NonEmptyArray<T>): T;
  * @param array The array
  * @returns A random element of the array.
  */
-export function randomElementOfArray<T>(array: Array<T>): T | undefined;
+export function randomElementOfArray<T>(array: ReadonlyArray<T>): T | undefined;
 
-export function randomElementOfArray<T>(array: Array<T>): T | undefined {
+export function randomElementOfArray<T>(array: ReadonlyArray<T>): T | undefined {
 	return array[randomInt(array.length)];
 }