+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Finish translating common commands.
+
 ### Changed
 - `/sr` now presents song request embeds more quickly.
 - `/quo open` now logs the timestamp at which the queue was opened.