+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0] - 2023-04-02
 ### Added
 - Complete translations for common commands.
 
@@ -418,7 +418,7 @@ After updating, be sure to run `npm ci && npm run build:clean && npm run migrate
 ### Added
 - Initial commit
 
-[Unreleased]: https://github.com/AverageHelper/Gamgee/compare/v2.0.6...HEAD
+[2.1.0]: https://github.com/AverageHelper/Gamgee/compare/v2.0.6...v2.1.0
 [2.0.6]: https://github.com/AverageHelper/Gamgee/compare/v2.0.5...v2.0.6
 [2.0.5]: https://github.com/AverageHelper/Gamgee/compare/v2.0.4...v2.0.5
 [2.0.4]: https://github.com/AverageHelper/Gamgee/compare/v2.0.3...v2.0.4